+++ src/screens/transaction_details/components/operations/components/desktop/index.tsx
@@ -13,7 +13,9 @@ import {
 } from '@material-ui/core';
 import { AvatarName } from '@components';
 import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
-import { TOKEN_DETAILS } from '@utils/go_to_page';
+import {
+  TOKEN_DETAILS, NFT_DETAILS,
+} from '@utils/go_to_page';
 import { columns } from './utils';
 import { OperationType } from '../../../../types';
 
@@ -21,6 +23,14 @@ const Desktop: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =
   const { t } = useTranslation('transactions');
   const formattedItems = props.items.map((x) => {
     const isToken = R.pathOr('', ['identifier'], x).split('-').length === 2;
+    const isNft = R.pathOr('', ['identifier'], x).split('-').length === 3;
+    let link;
+    if (isToken) {
+      link = TOKEN_DETAILS;
+    }
+    if (isNft) {
+      link = NFT_DETAILS;
+    }
 
     return ({
       action: x.action.replace(/([A-Z])/g, ' $1').toUpperCase(),
@@ -42,13 +52,13 @@ const Desktop: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =
         />
       ),
       value: (
-        isToken ? (
+        link ? (
           <div>
             <Typography component="span">
               {formatNumber(x.value.value, x.value.exponent)}
               {' '}
             </Typography>
-            <Link href={TOKEN_DETAILS(x.identifier)} passHref>
+            <Link href={link(x.identifier)} passHref>
               <Typography component="a">
                 {x.value.displayDenom.toUpperCase()}
               </Typography>